LayoutTests/platform/glib/imported/w3c/web-platform-tests/service-workers/service-worker/opaque-response-preloaded.https-expected.txt

CONSOLE MESSAGE: Response served by service worker is opaque
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/simple.txt due to access control checks.
CONSOLE MESSAGE: Response served by service worker is opaque
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/simple.txt due to access control checks.

PASS Opaque responses should not be reused for XHRs, loading case
PASS Opaque responses should not be reused for XHRs, done case
